Minor fixes

+++ config/epicollect/setup.php
@@ -42,9 +42,9 @@ return [
             'upload' => env('RESPONSE_DELAY_UPLOAD_REQUEST', 500000000)
         ],
         'rate_limit_per_minute' => [
-            'media' => env('API_RATE_LIMIT_MEDIA', 30),
-            'entries' => env('API_RATE_LIMIT_ENTRIES', 60),
-            'project' => env('API_RATE_LIMIT_PROJECT', 60),
+            'media' => (int) env('API_RATE_LIMIT_MEDIA', 30),
+            'entries' => (int) env('API_RATE_LIMIT_ENTRIES', 60),
+            'project' => (int) env('API_RATE_LIMIT_PROJECT', 60),
         ]
     ],
     'ldap' => [
